% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gen.R
\name{beta_gen}
\alias{beta_gen}
\title{Function to generate beta surfaces}
\usage{
beta_gen(surface = "ridge", len, surfplot = F)
}
\arguments{
\item{surface}{Choose either "ridge", "lagged", or "immediate".}

\item{len}{The length of the grids over which the surface is evaluated.}

\item{surfplot}{An indicator of whether or not to output a 3D-perspective plot of the surface. The default is False.}
}
\value{
Output is a matrix of values corresponding to the bivariate function evaluated over the gridpoints of t and v.
}
\description{
This function generates one of 3 beta surfaces examined in the simulations of Meyer, et. al. (2015): 1) ridge, 2) lagged, and 3) immediate.
The bivariate function is evaluated over grids t and v, assumed to be of equal length and with equal spacing. The length of the grid must be
specified. A 3D perspective plot of the generated surface may also be output.
}
\examples{
beta_gen("ridge",len = 225,F)
}
