% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelet_compress.R
\name{Wavelet_compress}
\alias{Wavelet_compress}
\title{Discrete Wavelet Transformation and Compression}
\usage{
Wavelet_compress(X_raw, wavespecs, alpha, xticks)
}
\arguments{
\item{X_raw}{The original observed matrix X of functions sample over a grid.}

\item{wavespecs}{A list containing the specifications for the DWT to be performed on X_raw.}

\item{alpha}{compression levels; i.e, P=0.95 represents a compression ratios of over 100:1, which is considered as compression levels. See more details in Morris et al,(2011).}

\item{xticks}{compression paraemters for X. See more details in Morris et al, (2011).}
}
\value{
Output from this function is a list containing 1) "compRes", 2) "compSet", 3)"D_all", 4)"C", 5) "keep", and 6) "wavespecsc".
CompRes: return compression result; CompSet: settings for compression on X; D: wavelet coefficients for X;  C: wavelet levels; Keep: an indicator vector with 1s and 0s.
1s mean we retain its correspoinding columns in the compression and 0s mean we exclude its correspoinding columns after the compression.
}
\description{
This function performs a DWT on each row (observed function) of X according to methods described in Morris and Carroll (2006).
In addition, performs joint compression on all of the rows of wavelet coefficients as done in Morris et al. (2011). Note that whereas in Meyer et al. (2015)
this function can accomdate multi-dimensional wavelet decomposition, our function only works for 1-dimension.
}
\examples{

}
