% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfmm_setup.R
\name{wfmm_setup}
\alias{wfmm_setup}
\title{Function to setup the design matrix}
\usage{
wfmm_setup(Y, X, Z, MCMCspecs, wavespecsx, pca, testspecs, pcaspecsx)
}
\arguments{
\item{Y}{N-by-T matrix of observed response functions. N is the total number of observed curves, T is the number of grid points at which the curves were observed.}

\item{X}{N-by-V matrix of observed predictor functions. N is the total number of observed curves, V is the number of grid points at which the curves were observed.}

\item{Z}{N-by-n random effects design matrix. N is the total number of observed curves, n is the number of subjects.}

\item{MCMCspecs}{A list of specifications for the MCMC algorithm used to estimate the beta coefficients. Needed for the wfmm executable file.}

\item{wavespecsx}{A list of specifications for the wavelet transformation of X. This list must include: "comp.level", a proportion supplied as the compression level, "alpha", "xticks", "nlevels", "wavelet", "wtmode", and "ndim"}

\item{pca}{An indicator of whether or not PCA is to be performed on X after the DWT. 1 indicates yes, 0 no. Default is 1.}

\item{testspecs}{A list of test specifications for posterior inference. This list must contain "twosample", an indicator of whether the design matrix has two samples, and FDR, a vector containing cutoffs for the Bayesian FDR procedure.}

\item{pcaspecsx}{If pca = 1, a list of specifications for the PCA performed on X. The list should contain "pca_level", the percentage of explained variation chosen as a cutoff to determine
how many PC's to keep, and "pca_cutoff", a vector of such pca level values to consider. If the argument is not supplied, defaults to
"pca_cutoff" = c(0.9, 0.95, 0.975, 0.99, 0.995, 0.999, 0.9999) and "pca_level" = 0.99.}
}
\value{
This function returns a list of 3 items: "model", "xspecs", and "testspecs". "model" contains "Z" (unchanged), "X" which is now the basis-space design matrix, and "C". "xspecs" is a list including the wavespecs and pcaspecs used for x. "testspecs" includes the original "testspecs" as well as splits the "FDR" vector into "alf" and "delt".
}
\description{
This function performs the specified basis operations on X (e.g. discrete wavelet transformation, PCA). This function should be used prior to using the wfmm executable file, which performs the BFOFR as described in Meyer, et. al. (2015)
}
\examples{

}
